requirements.txt
Flask==3.0.3
line-bot-sdk==3.17.1
python-dotenv==1.0.1
supabase==2.17.0
gunicorn==23.0.0
pandas==2.2.2
openpyxl==3.1.2